daemon: move datapath-related code to `datapath.go`
Signed-off by: Ian Vernon <ian@cilium.io>
